+++ .gitignore
@@ -1,2 +1,3 @@
 /cnc*
 __debug_bin*
+drill.yaml
